Refactored React Layout scaffolding pages for all UI
