+++ CHANGELOG.md
@@ -92,3 +92,6 @@
 
 ## 1.0.27 — 2025-09-27
 - Refactored React Chakra scaffolding and applied the Factory and Adapter pattern for multi-UI page generation
+
+## 1.0.28 — 2025-09-28
+- Refactored React Layout scaffolding pages for all UI
