+++ lib/layouts/adapters/antd.js
@@ -1,0 +1,7 @@
+const { source } = require("../antd/layout");
+
+function getLayoutSource(useTS) {
+  return source(useTS);
+}
+
+module.exports = { getLayoutSource };
