+++ lib/layouts/adapters/bootstrap.js
@@ -1,0 +1,7 @@
+const { source } = require("../bootstrap/layout");
+
+function getLayoutSource() {
+  return source();
+}
+
+module.exports = { getLayoutSource };
