+++ lib/layouts/adapters/chakra.js
@@ -1,0 +1,7 @@
+const { source } = require("../chakra/layout");
+
+function getLayoutSource(useTS) {
+  return source(useTS);
+}
+
+module.exports = { getLayoutSource };
