+++ lib/layouts/adapters/mui.js
@@ -1,0 +1,7 @@
+const { source } = require("../mui/layout");
+
+function getLayoutSource(useTS) {
+  return source(useTS);
+}
+
+module.exports = { getLayoutSource };
