+++ lib/layouts/adapters/tailwind.js
@@ -1,0 +1,7 @@
+const { source } = require("../tailwind/layout");
+
+function getLayoutSource() {
+  return source();
+}
+
+module.exports = { getLayoutSource };
