+++ lib/layouts/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  ...require("./factory/layout-factory"),
+};
