+++ lib/scaffold-react-project-structure.js
@@ -2,6 +2,7 @@
 const fs = require("fs");
 const path = require("path");
 const { createMUIPages, createBootstrapPages, createTailwindPages, createAntdPages, createChakraPages } = require("./pages/index");
+const { createLayoutComponent } = require("./layouts/index");
 
 function writeFileSafe(p, content) {
   fs.mkdirSync(path.dirname(p), { recursive: true });
@@ -89,7 +90,7 @@ export const api = axios.create({
   createReusableComponents(srcDir, ui, extx, ext, useTS);
 
   // 12) Layout component
-  createLayoutComponent(srcDir, ui, extx, ext, useTS);
+  createLayoutComponent(srcDir, ui, extx, useTS);
 
    // 12) Custom context functions
   createCustomContext(srcDir, extx, useTS);
@@ -1067,793 +1068,6 @@ export const LoadingSpinner: React.FC<LoadingSpinnerProps> = ({
 export default LoadingSpinner;`);
     }
   }
-}
-
-function createLayoutComponent(srcDir, ui, extx, ext, useTS) {
-  const layout = path.join(srcDir,"layout",`Layout.${extx}`);
-  if (!fs.existsSync(layout)) {
-    if (ui === "mui") {
-      writeFileSafe(layout, `import React, { useState } from 'react';
-import { Outlet, useNavigate, useLocation } from 'react-router-dom';
-import {
-  AppBar,
-  Toolbar,
-  Typography,
-  Box,
-  Container,
-  Drawer,
-  List,
-  ListItem,
-  ListItemIcon,
-  ListItemText,
-  IconButton,
-  useTheme,
-  useMediaQuery,
-  Menu,
-  MenuItem,
-  Chip,
-  ListItemButton,
-  Select,
-  ${useTS ? 'type SelectChangeEvent' : ''}
-} from '@mui/material';
-import {
-  Menu as MenuIcon,
-  Home,
-  Info,
-  Dashboard,
-  Person,
-  Notifications,
-  AccountCircle
-} from '@mui/icons-material';
-
-import i18n from '../i18n';
-import { useLanguage } from '../hooks/useLanguage';
-import { languages } from '../utils/common';
-
-const Layout = () => {
-  const [mobileOpen, setMobileOpen] = useState(false);
-  ${useTS ? "const [anchorEl, setAnchorEl] = useState<HTMLElement | null>(null);": "const [anchorEl, setAnchorEl] = useState(null);"}
-  const navigate = useNavigate();
-  const location = useLocation();
-  const theme = useTheme();
-  const isMobile = useMediaQuery(theme.breakpoints.down('md'));
-
-  const { currentLang, setCurrentLang } = useLanguage();
-
-  const menuItems = [
-    { key: 'home', icon: <Home />, path: '/' },
-    { key: 'about', icon: <Info />, path: '/about' },
-    { key: 'dashboard', icon: <Dashboard />, path: '/dashboard' },
-    { key: 'profile', icon: <Person />, path: '/profile' },
-  ];
-
-  const handleDrawerToggle = () => setMobileOpen(!mobileOpen);
-  const handleProfileMenuOpen = (event ${useTS ? ': React.MouseEvent<HTMLButtonElement>' : ''}) => {
-    setAnchorEl(event.currentTarget);
-  };
-  const handleProfileMenuClose = () => setAnchorEl(null);
-
-  const handleLanguageChange = (event ${useTS ? ': SelectChangeEvent' : ''} ) => {
-    setCurrentLang(event.target.value);
-  };
-
-  const drawer = (
-    <Box sx={{ width: 250 }}>
-      <Box sx={{ p: 2, borderBottom: 1, borderColor: 'divider' }}>
-        <Typography variant="h6" color="primary" fontWeight="bold">
-          PolyFront
-        </Typography>
-        <Typography variant="caption" color="text.secondary">
-          Multi-Frontend Scaffold
-        </Typography>
-      </Box>
-      <List>
-        {menuItems.map((item) => (
-          <ListItem disablePadding key={item.key}>
-            <ListItemButton
-              onClick={() => {
-                navigate(item.path);
-                if (isMobile) setMobileOpen(false);
-              }}
-              selected={location.pathname === item.path}
-              sx={{
-                cursor: 'pointer',
-                '&.Mui-selected': {
-                  backgroundColor: 'primary.main',
-                  color: 'primary.contrastText',
-                  '&:hover': {
-                    backgroundColor: 'primary.dark',
-                  },
-                },
-              }}
-            >
-              <ListItemIcon sx={{ color: 'inherit' }}>
-                {item.icon}
-              </ListItemIcon>
-              <ListItemText primary={i18n.t(\`common.\${item.key}\`, currentLang)} />
-            </ListItemButton>
-          </ListItem>
-        ))}
-      </List>
-    </Box>
-  );
-
-  return (
-    <Box sx={{ display: 'flex', minHeight: '100vh' }}>
-      <AppBar position="fixed" sx={{ zIndex: theme.zIndex.drawer + 1 }}>
-        <Toolbar>
-          <IconButton
-            color="inherit"
-            aria-label="open drawer"
-            edge="start"
-            onClick={handleDrawerToggle}
-            sx={{ mr: 2, display: { md: 'none' } }}
-          >
-            <MenuIcon />
-          </IconButton>
-          
-          <Typography variant="h6" component="div" sx={{ flexGrow: 1 }}>
-            PolyFront
-          </Typography>
-
-          <Box sx={{ display: 'flex', alignItems: 'center', gap: 1 }}>
-            <Chip label="v1.0.27" size="small" variant="outlined" />
-            {/* Language selector using i18n */}
-            <Select
-              value={currentLang}
-              onChange={handleLanguageChange}
-              size="small"
-              sx={{
-                color: 'white',
-                borderColor: 'white',
-                '.MuiSvgIcon-root': { color: 'white' },
-              }}
-            >
-              {languages.map((lang) => (
-                <MenuItem key={lang.code} value={lang.code}>
-                  {lang.label}
-                </MenuItem>
-              ))}
-            </Select>
-            <IconButton color="inherit">
-              <Notifications />
-            </IconButton>
-            <IconButton
-              color="inherit"
-              onClick={handleProfileMenuOpen}
-            >
-              <AccountCircle />
-            </IconButton>
-          </Box>
-        </Toolbar>
-      </AppBar>
-
-      <Box
-        component="nav"
-        sx={{ width: { md: 250 }, flexShrink: { md: 0 } }}
-      >
-        <Drawer
-          variant="temporary"
-          open={mobileOpen}
-          onClose={handleDrawerToggle}
-          ModalProps={{ keepMounted: true }}
-          sx={{
-            display: { xs: 'block', md: 'none' },
-            '& .MuiDrawer-paper': { boxSizing: 'border-box', width: 250 },
-          }}
-        >
-          {drawer}
-        </Drawer>
-        <Drawer
-          variant="permanent"
-          sx={{
-            display: { xs: 'none', md: 'block' },
-            '& .MuiDrawer-paper': { boxSizing: 'border-box', width: 250 },
-          }}
-          open
-        >
-          {drawer}
-        </Drawer>
-      </Box>
-
-      <Box
-        component="main"
-        sx={{
-          flexGrow: 1,
-          p: 3,
-          width: { md: \`calc(100% - 250px)\` },
-          mt: 8,
-        }}
-      >
-        <Container maxWidth="xl">
-          <Outlet />
-        </Container>
-      </Box>
-
-      <Menu
-        anchorEl={anchorEl}
-        open={Boolean(anchorEl)}
-        onClose={handleProfileMenuClose}
-      >
-        <MenuItem onClick={handleProfileMenuClose}>Profile</MenuItem>
-        <MenuItem onClick={handleProfileMenuClose}>Settings</MenuItem>
-        <MenuItem onClick={handleProfileMenuClose}>Logout</MenuItem>
-      </Menu>
-    </Box>
-  );
-};
-
-export default Layout;`);
-    } else if (ui === "bootstrap") {
-      writeFileSafe(layout, `import { useState } from 'react';
-import { Outlet, useNavigate, useLocation } from 'react-router-dom';
-import i18n from '../i18n';
-import { useLanguage } from '../hooks/useLanguage';
-import { languages } from '../utils/common';
-
-
-const Layout = () => {
-  const [mobileOpen, setMobileOpen] = useState(false);
-  const navigate = useNavigate();
-  const location = useLocation();
-  const { currentLang, setCurrentLang } = useLanguage();
-
-  const menuItems = [
-    { key: 'home', path: '/' },
-    { key: 'about', path: '/about' },
-    { key: 'dashboard', path: '/dashboard' },
-    { key: 'profile', path: '/profile' },
-  ];
-
-  const handleDrawerToggle = () => setMobileOpen(!mobileOpen);
-
-  return (
-    <div className="d-flex min-vh-100">
-      {/* Sidebar */}
-      <nav className="bg-dark text-white" style={{ width: '250px' }}>
-        <div className="p-3 border-bottom">
-          <h5 className="text-primary fw-bold mb-1">PolyFront</h5>
-          <small className="text-muted">{i18n.t('common.welcome', currentLang)}</small>
-        </div>
-        <ul className="nav flex-column">
-          {menuItems.map((item) => (
-            <li className="nav-item" key={item.key}>
-              <button
-                className={\`nav-link text-white w-100 text-start border-0 bg-transparent \${location.pathname === item.path ? 'active bg-primary' : ''}\`}
-                onClick={() => navigate(item.path)}
-              >
-                {i18n.t(\`common.\${item.key}\`, currentLang)}
-              </button>
-            </li>
-          ))}
-        </ul>
-      </nav>
-
-      {/* Main content */}
-      <div className="flex-grow-1">
-        {/* Header */}
-        <header className="bg-white border-bottom p-3">
-          <div className="d-flex justify-content-between align-items-center">
-            <button
-              className="btn btn-outline-secondary d-md-none"
-              onClick={handleDrawerToggle}
-            >
-              ☰
-            </button>
-            <h4 className="mb-0">PolyFront</h4>
-            <div className="d-flex align-items-center gap-2">
-              <span className="badge bg-secondary">v1.0.27</span>
-              
-              {/* Language Selector */}
-              <select
-                title='Select Language'
-                className="form-select form-select-sm"
-                value={currentLang}
-                onChange={(e) => setCurrentLang(e.target.value)}
-              >
-                {languages.map((lang) => (
-                  <option key={lang.code} value={lang.code}>
-                    {lang.label}
-                  </option>
-                ))}
-              </select>
-
-              <button className="btn btn-outline-secondary btn-sm">👤</button>
-            </div>
-          </div>
-        </header>
-
-        {/* Page content */}
-        <main className="p-4">
-          <Outlet />
-        </main>
-      </div>
-    </div>
-  );
-};
-
-export default Layout;`);
-    } else if (ui === "tailwind") {
-      writeFileSafe(layout, `import { useState } from 'react';
-      import { Outlet, useNavigate, useLocation } from 'react-router-dom';
-      import i18n from '../i18n';
-      import { useLanguage } from '../hooks/useLanguage';
-      import { languages } from '../utils/common';
-      
-      const Layout = () => {
-        const navigate = useNavigate();
-        const location = useLocation();
-      
-        // Language from context
-        const { currentLang, setCurrentLang } = useLanguage();
-      
-        const menuItems = [
-          { key: 'home', path: '/' },
-          { key: 'about', path: '/about' },
-          { key: 'dashboard', path: '/dashboard' },
-          { key: 'profile', path: '/profile' },
-        ];
-      
-        const [mobileOpen, setMobileOpen] = useState(false);
-        const handleDrawerToggle = () => setMobileOpen(!mobileOpen);
-      
-        return (
-          <div className="flex min-h-screen">
-            {/* Sidebar */}
-            <nav className="bg-gray-900 text-white w-64 hidden md:block">
-              <div className="p-6 border-b border-gray-700">
-                <h5 className="text-blue-400 font-bold text-lg mb-1">PolyFront</h5>
-                <p className="text-gray-400 text-sm">{i18n.t('common.welcome', currentLang)}</p>
-              </div>
-              <ul className="py-4">
-                {menuItems.map((item) => (
-                  <li key={item.key}>
-                    <button
-                      className={\`w-full text-left px-6 py-3 text-gray-300 hover:bg-gray-800 hover:text-white transition-colors \${
-                        location.pathname === item.path ? 'bg-blue-600 text-white' : ''
-                      }\`}
-                      onClick={() => navigate(item.path)}
-                    >
-                      {i18n.t(\`common.\${item.key}\`, currentLang)}
-                    </button>
-                  </li>
-                ))}
-              </ul>
-            </nav>
-      
-            {/* Main content */}
-            <div className="flex-1">
-              {/* Header */}
-              <header className="bg-white border-b border-gray-200 p-4 flex justify-between items-center">
-                <button
-                  className="md:hidden bg-gray-200 p-2 rounded-lg"
-                  onClick={handleDrawerToggle}
-                >
-                  ☰
-                </button>
-                <h4 className="text-xl font-semibold text-gray-900">PolyFront</h4>
-      
-                {/* Version + Language + User */}
-                <div className="flex items-center gap-3">
-                  {/* Version */}
-                  <span className="bg-gray-100 text-gray-700 px-2 py-1 rounded-full text-sm">v1.0.27</span>
-      
-                  {/* Language selector */}
-                  <select
-                    title='Select Language'
-                    value={currentLang}
-                    onChange={(e) => setCurrentLang(e.target.value)}
-                    className="bg-gray-100 text-gray-700 px-2 py-1 rounded-full text-sm"
-                  >
-                    {languages.map((lang) => (
-                      <option key={lang.code} value={lang.code}>
-                        {lang.label}
-                      </option>
-                    ))}
-                  </select>
-      
-                  {/* User button */}
-                  <button className="bg-gray-200 p-2 rounded-lg hover:bg-gray-300 transition-colors">
-                    👤
-                  </button>
-                </div>
-              </header>
-      
-              {/* Page content */}
-              <main className="p-6">
-                <Outlet />
-              </main>
-            </div>
-          </div>
-        );
-      };
-      
-      export default Layout;`);
-    }else if (ui === "antd") {
-      writeFileSafe(layout, `import { useState } from 'react';
-import { Outlet, useNavigate, useLocation } from 'react-router-dom';
-import { Layout as AntLayout, Menu as AntMenu, Drawer, Button, Typography, Select, Tag, Badge, Dropdown, Grid } from 'antd';
-import { MenuOutlined, HomeOutlined, InfoCircleOutlined, DashboardOutlined, UserOutlined, BellOutlined } from '@ant-design/icons';
-${useTS?'import type { MenuProps } from \'antd\';':''} 
-import i18n from '../i18n';
-import { useLanguage } from '../hooks/useLanguage';
-import { languages } from '../utils/common';
-const { Header, Sider, Content } = AntLayout;
-const Layout = () => {
-const [mobileOpen, setMobileOpen] = useState(false);
-const navigate = useNavigate();
-const location = useLocation();
-const screens = Grid.useBreakpoint();
-const { currentLang, setCurrentLang } = useLanguage();
-const menuItems = [
-{ key: '/', icon: <HomeOutlined />, label: i18n.t('common.home', currentLang) },
-{ key: '/about', icon: <InfoCircleOutlined />, label: i18n.t('common.about', currentLang) },
-{ key: '/dashboard', icon: <DashboardOutlined />, label: i18n.t('common.dashboard', currentLang) },
-{ key: '/profile', icon: <UserOutlined />, label: i18n.t('common.profile', currentLang) },
-];
-const profileMenu = {
-items: [
-{ key: 'profile', label: 'Profile' },
-{ key: 'settings', label: 'Settings' },
-{ type: 'divider' },
-{ key: 'logout', label: 'Logout' },
-] ${useTS?'as MenuProps[\'items\'],':''} 
-};
-const handleLanguageChange = (value ${useTS ? ': string' : ''}) => {
-setCurrentLang(value);
-};
-const Sidebar = (
-<div style={{ width: 250 }}>
-<div style={{ padding: 16, borderBottom: '1px solid #f0f0f0' }}>
-<Typography.Title level={5} style={{ margin: 0, color: 'var(--ant-color-primary)' }}>PolyFront</Typography.Title>
-<Typography.Text type="secondary">Multi-Frontend Scaffold</Typography.Text>
-</div>
-<AntMenu
-mode="inline"
-selectedKeys={[location.pathname]}
-onClick={({ key }) => { navigate(key); if (!screens.md) setMobileOpen(false); }}
-items={menuItems}
-style={{ borderRight: 0 }}
-/>
-</div>
-);
-  return (
-    <AntLayout style={{ minHeight: "100vh" }}>
-      <Header
-        style={{
-          position: "fixed",
-          insetInlineStart: 0,
-          insetInlineEnd: 0,
-          zIndex: 1000,
-          width: "100%",
-          display: "flex",
-          alignItems: "center",
-          gap: 12,
-        }}
-      >
-        {!screens.md && (
-          <Button
-            type="text"
-            icon={<MenuOutlined />}
-            aria-label="Open navigation"
-            onClick={() => setMobileOpen(true)}
-          />
-        )}
-        <Typography.Title
-          level={4}
-          style={{ color: "white", margin: 0, flex: 1 }}
-        >
-          PolyFront
-        </Typography.Title>
-        <Tag bordered={false}>v1.0.27</Tag>
-        <Select
-          size="small"
-          value={currentLang}
-          onChange={handleLanguageChange}
-          options={languages.map((l) => ({ value: l.code, label: l.label }))}
-          style={{ width: 120 }}
-        />
-        <Badge dot>
-          <Button
-            type="text"
-            icon={<BellOutlined />}
-            aria-label="Notifications"
-          />
-        </Badge>
-        <Dropdown menu={profileMenu} trigger={["click"]}>
-          <Button type="text" icon={<UserOutlined />} aria-label="Account" />
-        </Dropdown>
-      </Header>
-      <AntLayout>
-        {screens.md ? (
-          <Sider
-            width={250}
-            style={{
-              marginTop: 64,
-              background: "var(--ant-layout-sider-background)",
-            }}
-          >
-            {Sidebar}
-          </Sider>
-        ) : (
-          <Drawer
-            open={mobileOpen}
-            onClose={() => setMobileOpen(false)}
-            width={250}
-            bodyStyle={{ padding: 0 }}
-          >
-            {Sidebar}
-          </Drawer>
-        )}
-        <AntLayout style={{ marginTop: 64 }}>
-          <Content style={{ padding: 24 }}>
-            <Outlet />
-          </Content>
-        </AntLayout>
-      </AntLayout>
-    </AntLayout>
-  );
-};
-
-export default Layout;`);
-    }else if (ui === "chakra"){
-       writeFileSafe(layout, `import React from 'react';
-import { Outlet, useNavigate, useLocation } from 'react-router-dom';
-import {
-  Box,
-  Flex,
-  IconButton,
-  useDisclosure,
-  VStack,
-  HStack,
-  Text,
-  Container,
-  Badge,
-  Menu,
-  Button,
-  Separator,
-  Avatar,
-  Drawer,
-  Select,
-  Portal,
-  createListCollection,
-} from '@chakra-ui/react';
-import { useColorMode, useColorModeValue } from '../hooks/useColorMode';
-import { FiMenu, FiHome, FiInfo, FiGrid, FiUser, FiBell, FiChevronDown, FiSun, FiMoon } from 'react-icons/fi';
-
-import i18n from '../i18n';
-import { useLanguage } from '../hooks/useLanguage';
-import { languages } from '../utils/common';
-
-const sidebarWidth = 250;
-
-${useTS ? `type NavItemProps = {
-  icon: React.ReactNode;
-  label: string;
-  active?: boolean;
-  onClick?: () => void;
-};` : ''}
-
-function NavItem({ icon, label, active, onClick } ${useTS ?': NavItemProps': ''}) {
-  // Hover/background + text colors (fix: ensure text is visible on hover)
-  const hoverBg = active ? 'blue.600' : useColorModeValue('gray.100', 'gray.700');
-  const baseText = active ? 'white' : useColorModeValue('gray.800', 'gray.100');
-  const hoverText = active ? 'white' : useColorModeValue('black', 'white');
-
-  return (
-    <HStack
-      gap={3}
-      px={3}
-      py={2.5}
-      borderRadius="md"
-      cursor="pointer"
-      bg={active ? 'blue.500' : 'transparent'}
-      color={baseText}
-      _hover={{ bg: hoverBg, color: hoverText }}
-      onClick={onClick}
-    >
-      <Box fontSize="lg">{icon}</Box>
-      <Text>{label}</Text>
-    </HStack>
-  );
-}
-
-function Layout() {
-  // v3 disclosure: { open, setOpen, onOpen, onClose }
-  const { open, setOpen, onOpen, onClose } = useDisclosure();
-  const navigate = useNavigate();
-  const location = useLocation();
-  const { colorMode, toggleColorMode } = useColorMode();
-
-  const headerBg = useColorModeValue('blue.600', 'blue.700');
-  const pageBg = useColorModeValue('gray.50', 'gray.900');
-  const sidebarBg = useColorModeValue('white', 'gray.800');
-
-  const { currentLang, setCurrentLang } = useLanguage();
-
-  const menuItems = [
-    { key: 'home', icon: <FiHome />, path: '/' },
-    { key: 'about', icon: <FiInfo />, path: '/about' },
-    { key: 'dashboard', icon: <FiGrid />, path: '/dashboard' },
-    { key: 'profile', icon: <FiUser />, path: '/profile' },
-  ];
-
-  // Select (v3) expects a collection + string[] value
-  const langCollection = createListCollection({
-    items: languages.map((l) => ({ label: l.label, value: l.code })),
-  });
-
-  const drawerContent = (
-    <VStack align="stretch" gap={1}>
-      <Box px={2} py={3}>
-        <Text fontWeight="bold" color="blue.500">
-          PolyFront
-        </Text>
-        <Text fontSize="xs" color="gray">
-          Multi-Frontend Scaffold
-        </Text>
-      </Box>
-      <Separator />
-      {menuItems.map((item) => (
-        <NavItem
-          key={item.key}
-          icon={item.icon}
-          label={i18n.t(\`common.\${item.key}\`, currentLang)}
-          active={location.pathname === item.path}
-          onClick={() => {
-            navigate(item.path);
-            onClose();
-          }}
-        />
-      ))}
-    </VStack>
-  );
-
-  return (
-    <Flex minH="100vh" bg={pageBg}>
-      {/* Header */}
-      <Box
-        as="header"
-        position="fixed"
-        top={0}
-        left={0}
-        right={0}
-        zIndex={1000}
-        bg={headerBg}
-        color="white"
-        boxShadow="sm"
-      >
-        <Flex h={14} align="center" px={{ base: 3, md: 4 }}>
-          <IconButton
-            aria-label="Open menu"
-            variant="ghost"
-            color="white"
-            display={{ base: 'inline-flex', md: 'none' }}
-            onClick={onOpen}
-            mr={2}
-          >
-            <FiMenu />
-          </IconButton>
-
-          <Text fontWeight="bold">PolyFront</Text>
-          <Box flex="1" />
-
-          <HStack gap={2}>
-            <Badge>v1.0.27</Badge>
-
-            {/* Language Select (v3 controlled) */}
-            <Select.Root
-              size="sm"
-              width="160px"
-              collection={langCollection}
-              value={[currentLang]}
-              onValueChange={(e) => setCurrentLang(e.value[0] ?? currentLang)}
-            >
-              <Select.HiddenSelect />
-              <Select.Control
-                bg="transparent"
-                borderColor="whiteAlpha.700"
-                color="white"
-                _focusVisible={{ borderColor: 'white' }}
-              >
-                <Select.Trigger>
-                  <Select.ValueText placeholder="Language" />
-                </Select.Trigger>
-                <Select.IndicatorGroup>
-                  <Select.Indicator />
-                </Select.IndicatorGroup>
-              </Select.Control>
-
-              <Portal>
-                <Select.Positioner>
-                  <Select.Content>
-                    {langCollection.items.map((item) => (
-                      <Select.Item key={item.value} item={item}>
-                        {item.label}
-                        <Select.ItemIndicator />
-                      </Select.Item>
-                    ))}
-                  </Select.Content>
-                </Select.Positioner>
-              </Portal>
-            </Select.Root>
-
-            <IconButton aria-label="Notifications" variant="ghost" color="white">
-              <FiBell />
-            </IconButton>
-
-            {/* Account menu (v3) */}
-            <Menu.Root positioning={{ placement: 'bottom-end' }}>
-              <Menu.Trigger asChild>
-                <Button variant="ghost" color="white">
-                  <HStack gap={2}>
-                    <Avatar.Root size="sm">
-                      <Avatar.Fallback>U</Avatar.Fallback>
-                    </Avatar.Root>
-                    <Text display={{ base: 'none', md: 'inline' }}>Account</Text>
-                    <FiChevronDown />
-                  </HStack>
-                </Button>
-              </Menu.Trigger>
-              <Menu.Positioner>
-                <Menu.Content>
-                  <Menu.Item value="profile">Profile</Menu.Item>
-                  <Menu.Item value="settings">Settings</Menu.Item>
-                  <Menu.Item value="logout">Logout</Menu.Item>
-                </Menu.Content>
-              </Menu.Positioner>
-            </Menu.Root>
-
-            <IconButton aria-label="Toggle color mode" onClick={toggleColorMode} variant="ghost" color="white">
-              {colorMode === 'light' ? <FiMoon /> : <FiSun />}
-            </IconButton>
-          </HStack>
-        </Flex>
-      </Box>
-
-      {/* Sidebar (desktop) */}
-      <Box
-        as="nav"
-        display={{ base: 'none', md: 'block' }}
-        w={sidebarWidth}
-        pt={16}
-        px={3}
-        position="fixed"
-        h="100vh"
-        overflowY="auto"
-        borderRightWidth="1px"
-        bg={sidebarBg}
-      >
-        {drawerContent}
-      </Box>
-
-      {/* Drawer (mobile, v3) */}
-      <Drawer.Root open={open} onOpenChange={(e) => setOpen(e.open)} placement="start">
-        <Drawer.Backdrop />
-        <Drawer.Positioner>
-          <Drawer.Content>
-            <Drawer.CloseTrigger />
-            <Drawer.Header>PolyFront</Drawer.Header>
-            <Drawer.Body>{drawerContent}</Drawer.Body>
-          </Drawer.Content>
-        </Drawer.Positioner>
-      </Drawer.Root>
-
-      {/* Main */}
-      <Box flex="1" pl={{ base: 0, md: sidebarWidth }} pt={16} w="full">
-        <Container maxW="6xl" py={6}>
-          <Outlet />
-        </Container>
-      </Box>
-    </Flex>
-  );
-}
-
-export default Layout;
-`);
-
-    }
-}
 }
 
 function createCustomContext(srcDir, extx, useTS) {
